+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 Please clear changelog after each release.
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
-- Updated embedded FrozenLib.
-- Fixed a broken zip file.
+- Added unique sounds for Maple Leaves.